+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file contains all the latest changes and updates to Postal.
 
+## [Unreleased]
+
+### Features
+
+* **api:** add support for custom Message-ID in send/message API endpoint
+
 ## [3.14.0](https://github.com/MissionInbox/postal/compare/3.13.0...3.14.0) (2025-04-21)